refactor: Move fallback to separate file and reuse scaladoc symbol guessing
Turns out Scaladoc had a similar functionality, though I improved it a tiny bit and moved it to Symbol class.
